monitor/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
httpx==0.28.1
python-dotenv==1.0.1
pydantic==2.10.4
anthropic>=0.39.0
